Add tensegrity
I wrote this last night. The strikethroughs are transcriptions of what I
have on paper, and I decided to keep them in as part of the poem. I
kind of want to spend more time finding better words with better sounds,
but I decided that this more imperfect piece was fitting.